+++ src/services/utils.service.tsx
@@ -82,3 +82,13 @@ export const sortRows = (rows, header, direction) => {
         return a[header] > b[header] ? -1 : 1;
     });
 };
+
+export const downloadXmlAsBpmn = (model) => {
+    const element = document.createElement("a");
+    const file = new Blob([model.xmlData], {type: 'text/xml'});
+    element.href = URL.createObjectURL(file);
+    element.download = model.type === 'bpmn' ? toKebabCase(extractBpmnProcessName(model.xmlData) || model.name) + '.bpmn' : toKebabCase(extractDmnTableName(model.xmlData) || model.name) + '.dmn';
+    document.body.appendChild(element); // Required for this to work in FireFox
+    element.click();
+    document.body.removeChild(element);
+};
